Revamp of the codebase
This isn't a use-ready revamp of the code, but I just removed the redundant parts in each of the scheduling algorithms (i.e. the averages and the other calculations + the heavily shared data structures and variables) and touched up any residue from these changes without testing if these changes were accurate or not. This revamp is probably best continued on a separate branch so that such a major revamp is separated from the main code for good practice. (I probably should've revamped on a separate branch anyways but oh well, live and learn)