Further clean-up (no testing)
Another rough clean-up, mainly on the scheduling algorithm classes, but there has still been no testing done to ensure that they are functioning correctly. After pulling this branch back to main I will continue tidying up and test further to make sure that at least the terminal version of the project is ready. (will move on to the GUI later this week)